tabbar

+++ SecondWind/Classes/CollectionControllers/SWPlayerViewController.m
@@ -88,7 +88,7 @@
     view = [[UIImageView alloc] initWithFrame:CGRectMake(0, 0, 200.0f, 200.0f)];
     view.backgroundColor = [UIColor lightGrayColor];
     UILabel *label = [[UILabel alloc] initWithFrame:view.bounds];
-    label.text = [NSString stringWithFormat:@"%i", index];
+    label.text = [NSString stringWithFormat:@"%lu", (unsigned long)index];
     label.backgroundColor = [UIColor clearColor];
     label.textAlignment = NSTextAlignmentCenter;
     label.font = [label.font fontWithSize:50];
